Remove pregenerated header files from musl 1.1
Some pregenerated header files exist in the src/ctype and src/locale
directories. They are generated by out-of-tree code found on GitHub,
musl-chartable-tools.

This causes a slight reordering of the early bootstrap, as the early
environment is quite spartan and has broken/missing tools to regenerate
these files. With an added rebuild everything works fine.